[zircon] [scripts] Add continuation to simplify tun/tap after reboot
To make fx run -N even more friendly, I added "&&" (continue with
the next command if the first command succeeds) and "\" (continue
command on the next line) so users can simply copy and paste the
helpful commands in a single copy-and-paste action.

I find I have to re-enable tun/tap after each reboot of my Linux
machine. Of course, I don't remember until I try the "fx run" command,
but "fx run" helpfully prints the suggested commands to do that:

  To use qemu with networking on Linux, configure tun/tap:
  sudo ip tuntap add dev qemu mode tap user richkadel
  sudo ifconfig qemu up

Without thinking about it, it might seem like you should be able to
copy and paste both lines, but this fails because the sudo in the first
line tries to use the second line as the password.

The change simplifies running the commands, and may be less confusing.

Test: Verified results locally (very minor change)
Change-Id: I45ebfda917c96ba5704f108119b65a06aaba0ae4