[fdio] Don't crash on exit when blocked in accept()
Previously, processes would crash after their atexit handle ran if the
process contained a thread that was blocked in accept. When blocked in
accept, the thread reserves a slot in the FD table for the socket that
will be accepted. Prior to this CL, the fini logic that unwinds the FD
table did not understand how to handle such reserved slots.

Test: exit_in_accept_test in fdio-test
Change-Id: I60d10e6a48980dabc274ba180732adadec470b8e