[ledger] Use PageWatcher in LazyToEager integration test.
Previously, we relied on SyncStateWatchers to determine a moment when two
instances of the same page have obtained each other's updates. This is
unreliable: in case of a multi-piece object, the download state can turn
from IN_PROGRESS to IDLE and back several times as individual pieces are
being fetched. We will now rely instead on PageWatcher receiving
updates to the page.

TEST=SyncIntegrationTest/SyncIntegrationTest.LazyToEagerTransition
LE-704 LE-742 #done

Change-Id: Ieeed1a5a81f76d1fceedf63ee1fb5c5ee71d11a3