[zircon][sdmmc] Use a callback instead of a virtual interrupt
Changed the SDIO interrupt to use a registered callback instead of a
virtual interrupt. Added a virtual interrupt to the SDIO core driver to
support interrupt multiplexing which will be introduced in a later
change. The core driver will have to use virtual interrupts as
the function devices may eventually be in separate devhosts.

Bug: ZX-3879
Test: Bluetooth works on Cleo
Test: runtests -t mtk-sdmmc-test
Test: runtests -t sdmmc-test
Change-Id: I999832162fd0e980e1b3a7fc4fce96d65e550540