[mediasession] Remove nondeterminism in tests.
Some tests depend on a count of packets sent to client without ACK
accumulating. They send an ACK on startup. In rare cases these ACKS
can arrive late and reset the server's count, resulting in different
behavior than the tests expect and creating a failure.

To fix this, this CL removes the initial ACK so its timing will not
influence the test outcome.

FLK-192 #done

TEST: fx run-test mediasession_tests
Change-Id: I3e8478c08b5843e7645c2876b288252bbf248ba9