[host tests] Delete old copy dance
The deleted copies and actions were done so that host tests could be run out of $root_build_dir/host_tests both on the bots and locally with fx run-host-tests. The bots no longer need this and fx now consults tests.json to figure out the location of the tests.

Bug: IN-819 #done.
Change-Id: I6cf1213e68cb5dcbb3e67588978a8b29f90ae167