[netstack3] Replace receive_arp_response to send generic ARP packets
The receive_arp_response test function is not generic enough for the
upcoming gratuitous ARP implementation. This is necessary to send both
request and response ARP packets without reimplementing the same logic
and assertions.

The solution to this was to rename receive_arp_response to
send_arp_packet, which allows for specifying the operation of the ARP
packet (request or response). The function arguments were moved around
to mirror receive_arp_packet.

Change-Id: I815ff437656cdddd0f51f67983df24a95f1e124e