[rust][async] Prevent hot loop on false Closed signal
Previously, it was possible for a hot-looping task to be
created as a result of a stale PEER_CLOSED signal arriving
for a new, still-open channel. This change prevents this
by ensuring that the `PEER_CLOSED` signal gets reset and
re-waited in the event that a read or write occurs without
resulting in PEER_CLOSED.

This change fixes the issue for zx::Fifo and zx::Channel,
and a followup change will be opened for zx::Socket.

Change-Id: I0220b084320158d20d624b75c4267b0f87e6687f