[zircon][gn] fidl_library() support for llcpp
This implements half the FIDL generation for LLCPP bindings.  It
can't be done fully the straightforward way until we can build the
fidlgen tool (Go).  Instead, this integrates it halfway by doing the
JSON generation part.  Then instead of running fidlgen, it produces
a build API module (JSON) describing how it would have run fidlgen,
and expects to find the fidlgen output files in the source tree.
Later integration will take that information and do the regeneration
and/or check that it doesn't need to be done, so that the
higher-level build can verify what the Zircon build alone can't yet.

Bug: BLD-442 #comment fidl_library() support for llcpp
Bug: BLD-353 #comment Switchback for FIDL LLCPP bindings before unification
Change-Id: Ied2e9972da1c1bb4957e29f98c721d96d8839d97