[appmgr] Kill realm when its job dies
When realm's job dies, that realm becomes useless because it cannot
create any more components aor realms inside of it, so better to kill it
and lets realm owner know that realm was killed so that they can exit
gracefully or create a new one.

This will also complete bug CF-706 because now we have checks in place
which ensure that appmgr doesn't crashdue to closed handles  when there
is a oom.

CF-706 #done

TEST=fx run-test appmgr_integration_tests

Change-Id: I4a06bc73d76fe44a075d63fc3f0f762a94a19a14