[fshost][minfs] Add fshost tests for reformatting, verifying filesystems
Additionally, some minor related changes:
- Remove unused "FormatMinfs" function.
- Plumb the "Should we check filesystems?" boolean through explicit
arguments instead of environment variables.
- Make minfs superblock validation more strict. This validates that
clients can't format a filesystem with N blocks, and either mount
of check the consistency of a filesystem with M blocks, for M < N.
This change, along with the fshost tests, acts as a regression for
a bug where the consistency checker only inspected a subsection of the
filesystem.

Change-Id: I15470e9529553526011c8fba2d8a6f5987b8e1a6