[build][ldso] Use libprefix for all instrumented variants
Also add an explicit option in environment() to enable a per-variant
libprefix in all variants of a toolchain (or none).  By default, each
variant that has the "instrumented" tag gets "$variant/" appended to its
${toolchain.libprefix} value.  Remove the "asan" variant's explicit
toolchain_vars.libprefix setting, which is now redundant.

The dynamic linker is compiled to request the libprefix from the loader
service, and all executables have the prefixed name in PT_INTERP so they
get the same variant's dynamic linker.

Bug: BLD-429 #done
Bug: ZX-3624 #done
Change-Id: Ia74bc8d449aa6ce3ded00df8940b057e790486f4