[sys] Start component_test_runner, a component runner for tests
This starts a component_test_runner which is an implementation of the
fuchsia.sys.Runner interface that can instantiate a component within an
environment configured for tests.  The specification of the test
environment and of the component under test is the component manifest
itself. This way a test author can configure different environments to
run a test in without having to modify the code for the component under
test.

This patch responds to incoming Runner requests and starts parsing the
test specification data, but doesn't yet actually attempt to create the
test environment or actually instantiate the component.

Change-Id: Ibc70a60202c026174268a109a376c072bd370a10