[debugger] Watchpoint integration test.
This is the stub of the test, that is not joined to the build. This is
because a lot of logic for making this test to work is required and the
CL is big enough at it is. The following CL will do the actual
watchpoint logic and hook up this test to run.

TEST=No functional change.

Change-Id: Idee83c1e65aa55552242d5679e0ff178b8151fcc