[vulkan] Improve hw support for VK_LAYER_FUCHSIA_imagepipe_swapchain
This doesn't cause a change in behavior. The code is only used for
examples and the change simply increase hw support and allows for
optimizations in future drivers.

- Determine memory type index instead of assuming 0.
- This is a memory allocation dedicated to a single image. Inform
  the driver of this. Required by goldfish ICD and useful for
  other drivers.

The result of this change is that examples that use this layer can
be used when running fuchsia on aemu.

DX-939 #comment

Test: fx shell present_view fuchsia-pkg://fuchsia.com/vkcube_on_scenic#meta/vkcube_on_scenic.cmx
Change-Id: Ib27140ea6c5871e34af01d4ea631e2f4b80fa600