[kernel][x86] Add write_msr_safe
write_msr_safe attempts to write a value to an MSR; if it encounters a #GP
fault, it returns a status with the value unwritten.

Like read_msr_safe, this should be executed with interrupts disabled; I
plan to change both of them soon to assert that we are either cpu-pinned
or have hard interrupts disabled.

Tested: See 4th CL in stack; adds 'k' commands to access read_msr/write_msr.

ZX-3359

Change-Id: I6f91818bf1ffc34f54e041ad17aca1317d240321