[scenic] Fix multiple present checks in BaseView
BaseView::PresentScene() expects single call in a vsync, but BaseView::Present()
can queue multiple calls, i.e. simplest_app example. For the following calls, it
is OK to skip as they will fall into the same vsync interval. This CL removes
the checks for that case.
However, this check would make sense for InvalidateScene() driven updates, so
I noted that we should bring it back when that is the only option.

SCN-1379 #done

Change-Id: Ie3e662248d82d480fe2f2d04a955f635b3eecf39