[linux_runner] Replace SyncPtrs with fit::promise.
This removes the use of FIDL Sync pointers from the linux_runner, using
fit::promises in their place. This removes a large chunk of the
blocking logic from the linux_runner, however the gRPC components are
still synchronous. Making these async will happen in a subsequent
change.

This change will make it simpler to add unit tests to this component
since we won't have to work around blocking logic.

Test: tiles_ctl add linux://chromium
Change-Id: Ifbed771d8764153607d4ea483ff2c69a75835d5a