[mdns] add v4/v6_multicast_address properties to config
This CL adds the ability to configure the multicast addresses that
mdns will use for communication. This is needed temporarily to allow
multiple mdns implementations to run side-by-side, given the current
limitations of the network stack.

Previously, the MdnsAddresses class had static methods that took a
port parameter, and the port number was distributed through the
object graph. This new code changes the MdnsAddresses methods into
instance methods, the instance carrying the port number and the
two multicast addresses. The config code initializes an instance
of MdnsAddresses, and a reference to that instance is distributed
through the object graph instead of the port number.

TEST: fx run-test mdns_tests

Change-Id: I2f877f25a9d977b5913b1b4e392031e296666aaf