[fidl][difl] Process declarations in order, tables support
This changes the tool to evaluate declarations based on the dependency
order calculated in the frontend rather than by type. This means that if
declaration A references declaration B then B will be evaluated before
A.

This adds some support for evaluating changes to tables and some more
tests for types.

Change-Id: Icbf5593a3847c8b05036ed3a850b4c79ce8d651c