[debugger] Exposed strip separation in BuildIDIndex
Actually-stripped and separated debug symbols are, we are told, a thing
that is coming. We need to actually be able to express this in our API.

.build-id folders already have this separation. As of this patch that
will be expressed correctly. The symbol downloader may not download both
halves of the binary so that will have to come in another patch.
Un-indexed folders, single files, and ids.txt are still treated as
sources of unstripped full binaries. At least ids.txt and possibly some
of the others mentioned are deprecated, so patches may or may not come.

Change-Id: I191b76089b29eac2b83c6cf30ab1b87d8404aa14