[devcoordinator] Fix buggy composite device behavior
If a component was a device with MUST_ISOLATE set on it, we would try to
bind to the non-proxied side of it, which doesn't speak the device
protocols.

This is a re-landing of I64eaccec5d51997f4374d0277b040a952f681667.  The
change was reverted due to flakiness on ASAN builds.  The flakiness was
fixed in I7c6256d5f2a0af8922e85423ccb27330687182e6.

Change-Id: I8736ff598a764ef2d70bf7792019e3a749ad1eec