[zircon][gn] Clean up default_variants list
Make the preference for clang on host explicit in the first selector.
Instead of using environment matching to get clang for efi, just rely
on exclude_variant_tags in efi to skip over gcc and reach clang as a
final fallback in the list.  This previously wasn't possible because
of environment_redirect() issues, but the recent revamps to that code
has removed those limitations.

This should make sure there is a valid default available in other
environments than efi and host that are incompatible with gcc, such as
fuzzer environments.

This uncovered a latent bug in the userboot/vdso embedding magic,
where the kernel selecting the vDSO binary to embed and userboot
selecting the vDSO binary to get symbols from would not always use
the same binary.  Now they share the target that finds the binary.

Bug: BLD-430
Test: no effect on the build in variants=["gcc"] or variants=["clang"]
Change-Id: Ida4b3b4c4e51809c887f9e598b784e932f684ebf