[debugger][x64] Watchpoint Debug Agent Implementation
This implements the hook of the DebugAgent with the zircon syscall,
effectivelly implementing watchpoints.

This includes an integration test that sets up a watchpoint and verifies
that the exception was correctly triggered.

The test is disabled for arm64, as this functionality is not implemented
for that architecture.

Also disabled a software breakpoint test for arm64, that flakes because
zircon doesn't flush the IC when setting the register state, causing a
thread to sometimes hit a breakpoint that has been removed.

TEST=fx run-test debug_agent_tests

Change-Id: I11f33e94b7aec68e23a5abdca2743be5be4516d6