[carnelian] Add a new framebuffer drawing features, examples and tests
Added a drawing example that works only with software framebuffer and
exercises the new Canvas features.

Added very simplistic implementations of fill_circle() and fill_roundrect(). These should
be good enough for recovery-mode UI and for working up the non-drawing portions of
Carnelian until Spinel is available.

Added a simple Canvas test to break the ice.

Change-Id: I557c60a3e7ea4098240f622017f367183c8fda76