[text] Add TextFieldMultiplexer and tests
The TextFieldMultiplexer connects multiple TextFieldRequestStreams to a
single TextFieldProxy. Since we'd like to simplify TextField
implementations by allowing them to assume there's only a single client,
we need this code to de-interlace transactions.

This technique does have a latency penalty — it means a transaction has
to reach CommitEdit before it gets forwarded. However, I think that this
could be improved in the future without changing the overall structure,
by forwarding the first transaction that comes in message-by-message,
and making just the second transaction queue. This could be combined
with timeouts that call AbortEdit on the TextFieldProxy and switch to
the second transaction queue to avoid the first holding a lock for too
long.

This code isn't actually in use yet; it will get used in a future CL.

Change-Id: I86eeedbe22a175d23e1382c0dbf40028a3577d4e