[e2e] add end to end test for the terminal product
The test verifies that after reboot, the screen is not
black. It could be used for more than one product, but is configured
for now to test the terminal product.

The test is in the directory src/end_to_end_tests. The directory
src/ is organized by areas, but end to end tests by definition span
areas, so they get their own area under src.

That the test is added to the terminal product for now only makes it
build when the build configuration is terminal.*. It doesn't cause
the test to be executed by a CI/CQ builder of that product. That requires
that the builder is configured to pick up tests with the label assigned
to the test, e2e-terminal.

How to execute the test on a linux workstation:

  fx set terminal.astro
  fx build
  fx reboot -r
  fx pave -1
  FUCHSIA_IPV4_ADDR=$(buildtools/linux-x64/dev_finder list) \
  FUCHSIA_SSH_KEY=${PWD}/.ssh/pkey \
  FUCHSIA_TEST_OUTDIR=${PWD}/outdir \
  ./out/default/host_x64/screen_is_not_black_test

This is the output:

  00:00 +0: screen_is_not_black_test the startup screen is not black
  Target device: 192.168.1.231
  Try 0 at starting sl4f.
  Running over ssh: run fuchsia-pkg://fuchsia.com/sl4f#meta/sl4f.cmx
  SL4F has started.
  Saw a screen that is not black.
  Running over ssh: killall sl4f.cmx
  00:05 +1: All tests passed!

INTK-1031 #comment This change adds the test to run in the new builder.

Change-Id: I9bfe3055f12fd3dc58f94d57f82d15eadbb3137a