Switch to more forcefull reboot with PowerStatePowerCycleSoft
We have seen Sata M2 devices that were invisible after reset.
PowerStateMasterBusReset seems to do the equivalent of hitting
the now defunct reset button motherboards used to have, which is
not the same than a power cycle.
PowerStatePowerCycleSoft is not a real power cycle either given
that the CPU (AMT), memory, and the NIC are still running, but
we are hopeful that the rest of the system is powered off as indicated
by the power button LED going off when we send that command.

As a result, we think this may help with Sata devices being gone
after being potentially being wedged by a bad driver and not fixed
by a simple reset.

Change-Id: I9364e6d6ded43d4ab57c3c7fd01125d504d2c9b6