[modular] RELAND: Fix basemgr to pass config override to sessionmgr
This is a reland of f7a77a9a951fde2c4eec95e1a642ed98a83c806e with the
following fix:
* The /config dir which appmgr provides to a component is not clonable,
but the /config/data dir is -- failing to clone means  /config isnt
propogated to sessionmgr, which fails a test in CI, which is good, but
not in CQ. A CQ test for this would involve faking out a /config/data,
but the testing facility for it doesn't exist today. The orginal CL
below is amended to clone /config/data instead of /config.

Original CL:

* Re-factor ModularConfigReader to parse config file during construction, and
 expose some getters instead. Have it accept a root directory to work off
 of, which makes it more testable. (i.e., fake out a directory and pass it to
 ModuleConfigReader). Provide a factory for using the incoming namespace.

* Have basemgr always pass config data to sessionmgr in its
 /config_override namespace, since appmgr enforces that /config is not
 overridable. sessionmgr uses ModularConfigReader, which prefers
 /config_override over /config when available.

* Fix modular_config_xdr to parse SessionShellMapEntry -- it wasn't
 parsing session shell before.

Test: modular_config_unittest tests that ModularConfigReader APIs
parse and return basemgr config given a config directory. Tests that session
shell URL is being parsed.

Change-Id: I737dbe897f6dc1362518bdab16dd06347ea30924