[libc] Add "noop" smoke test for libc startup
This is the most basic test that program loading, dynamic linking,
and libc startup paths are working.  Nothing else would be running
if this didn't work, not even the standalone core-tests from
userboot.  But now also build this test for all the instrumented
variants, to make sure each instrumented libc build actually works.

The ubsan variants were not being tested before now and in fact are
broken.  So those variants are disabled until ZX-3986 is resolved.
When they are re-enabled, this test will sure they don't regress.

Bug: ZX-3986 #comment noop-test built for all instrumented variants
Change-Id: Iefc1c03415dbe942e7587cc7c9b01f01e6da82c3