[fidlcat] prefer posix sh over bash during build
On a system where bash is not the default shell, the placement of the shbang
line meant this was invoked as a POSIX sh script, resulting in `[[` not being
found, causing a build error.

We're in general not taking dependencies on bash during the build, where this
seems to be invoked, so on the path to fixing this, I altered the script to
be POSIX sh compatible.

Change-Id: I86cf9ae112da4217c3e25402c1fe5c1b6d27dda6