[build][rust] use the stripped path to binaries
Many package rule invocations seem to have copy-pasta'd paths to their
binaries. This is leaning on the "stripping fixup" that finalize_manifests.py
does, adding build time and other side effects to the build. Use the
canonical path to the binaries instead, which is at the top of root_build_dir
and is already stripped.

Change-Id: I846085768206c042884dfaa0e6b0fdf780790d29