[debugger][experimental] Backtrace Caches
This CL hooks up an experimental feature that is not activated for
normal debugger use.

The main idea is that the debugger will store backtraces for relevant
exceptions (breakpoints, general exceptions, etc.), so that they can be
reviewed later.

This CL attemps to be minimal about this, but surfacing a feature to
the console requires a fair amount of moving parts (settings and verbs).

TEST=Unit test injecting an exception and verifying that the cache
     stored the stack trace correctly.

Change-Id: I2fab74ac517d7e31de73cce8048efe6dcefd1cac