[linux_runner] Refactor for unit tests.
This change sets up a simple sanity test that demonstrates the ability
to simulate a guest system interacting over vsock. To accomplish this we
introduce a new testing library for faking out the launching of guest
systems; providing an API to simulate guest interactions instead of
launching real virtual machines.

Test: fx run-test linux_runner_tests
Change-Id: Ifc63ddff4d57200cac17a3922a487526f332df7c