[modular][login mode] Allow runtime login mode override.
This changelist changes basemgr to now check for a runtime
login mode override from SetUI service and not use auto-login
guest mode if it conflicts with this value. The basemgr also
persists the compile time configuration to SetUi service if
no specification has been made yet.

This change also contains a test for fxr/276756.

SU-167 #done
TEST: fx run-test modular_login_tests

Change-Id: I7ae9f9ee6dcf973567b10f1574380d42c2b480a6