[guest] Add BUILD.gn for "biscotti_guest/third_party/protos".
Currently, users of the protos directly include the ".proto" files
directly in their build rules, with constructs such as:

  sources = [
    "../../third_party/protos/xxx.proto"
  ]

This isn't ideal, because in the general case circumvents visibility,
causes files to be built multiple times, and makes moves more difficult.

Add a native BUILD.gn file to the directory, and provide a build rule
for C++ bindings of the VM protos and C++ bindings for the vsh proto.

Change-Id: Icc9e911e66cf18582098849c3cd6b7a59a7c81d2