[pi][trace] Add tracing for priority inheritance events. (kernel)
Add a ktrace events which show a PI chain ripple related using flow
events.  The chain starts with the thread whose actions result in a
change of OwnedWaitQueue pressure.  Typically, this is a thread which
either blocks or unblocks in an owned wait queue, but it can also be
caused by a thread which changes the base priority of a thread which
is currently blocked in an OwnedWaitQueue.

When threads which are downstream from the initiating wait queue
experience an actual change in effective priority as a result of
pressure change, they log the details of the change (old/new
inherited/effective priority) and link these events using flow IDs.

If the change of wait queue pressure result in no change of effective
priority for any thread, no trace events will be generated.

Note: this CL only adds trace definitions and tracing behavior to the
kernel.  To make any use of these events, a follow up change which
implementes the trace importer behavior is needed.

Change-Id: Ic2d795c7723c9c0d7562b5b4022fee9ac514f23b