[debugger] Added valid options for SettingSchema
This is the first step to implementing enums and enum-lists for the
setting system.

This is not hooked yet to the SettingStore, so it's possible to add
strings even though the schema doesn't allow it. As no setting uses this
functionality, that will be added in another CL for making reviewing
easier.

TEST=Added SettingSchema unit test.

Change-Id: I8376d9266c2fb3591576c9b14b5404db140cbd0b