[pm serve] fix interpretation of -d/-repo
The -d flag was altered some time ago to allow pointing at the repository
serving directory rather than the repository root. A recent regression caused
the interpretation to change, however, there is also now a discrepancy
between the meaning of -repo and -d. We now interpret -d to mean the
repository serving directory, and -repo to mean the repository root. As SDK
consumers are presently using -d, this restores the prior behavior for that
flag, while making the -repo flag behavior consistent across other commands.

Change-Id: I50d7567194101d3cf2c6af1bf83a33bc481f4290