[fidlgen][rust][cpp] Fixes to constants handlings
Until now we relied on the format of constants in FIDL source being compatible
with the target language. In the case of floats for rust, we were lucky that
things worked and had to write floats as '160.0' for instance instead of
'160'.

This change reduces the reliance on the format in source, and prepares
rust and cpp bindings to dealing with resolved values only.

FIDL-605 #comment

Test: fx run-host-tests fidlgen_{golang,cpp,rust}_ir_test

Change-Id: Ie87857e8d2967166d794b9cd556edfbe4ba39db4