[fidl] [llcpp] Synchronous event handlers
Implements synchronous event handlers that are safer than the
ExpectFoo() versions found in Go.

Given the FIDL protocol

    protocol Foo {
        -> OnOpen();
        -> OnSomthing(int arg);
        A(int a) -> (int b);
    }

We can build an "event dispatcher vtable" like this:

    struct FooEventHandlers {
        fit::function<zx_status()> on_open;
        fit::function<zx_status(int arg)> on_something;
        fit::function<zx_status()> unknown;
    };

and have the client call:

    Foo::SyncClient client(std::move(channel));
    client.HandleEvents(FooEventHandlers {
        .on_open = [&] { ... },
        .on_something = [&] { ... }
    });

Doing so will not interfere with regular method call replies, since all
method calls in llcpp are sync for now. In addition, we force the caller
to handle all possibilities, so there is less chance of abuse/dropping
messages.

TEST: All llcpp tests involving events are rewritten to use this new
API.

FIDL-604 #done

Change-Id: Ib59b77f113ccb3d975dbff24f57b420b6690c12a