[fidlc] Expose resolved values in JSON IR
Up until now, we emitted the raw value as it appears in .fidl source
files into the JSON IR. With the support of binary literals, backends
for target languages which did not support that syntax (e.g. Dart)
would have to resolve an interpret constants. Since we're already doing
this work in fidlc, we're simply emitting the resolved value, along with
the original value.

FIDL-486 #comment
FIDL-605 #done

Change-Id: I7e6bb28a4392f0a43be59f236a3222bb662ebb96