[ddk] Prevent instructions that aren't supported by the VMM's decoder
Compiling zircon with clang causes the guest integration tests to fail
because it emits instructions (that trap to the hypervisor) that both
read and write memory such as "and" and "or". This temporarily prevents
the compiler from using those instructions until the hypervisor's
instruction decoder supports them.

TEST=fx run-test guest_integration_tests

MAC-250 #done
MAC-251 #comment

Change-Id: I0f8779a6ffdbf3d75690112e3a2304aa95a800a6