[x86][hypervisor] Save and restore superset of extended registers
Changes the existing extneded register save and restore logic to always save and restore
the full set of potential states and not states that the guest currently has enabled.

Currently the host can be made to panic if the guest enables a state, and then later
disables it. Sequence of guest operations for the panic is added to the hypervisor
unit test.

This panic occurs as when xsave occurs any state present in xcr0 is saved. If xcr0 is
then reduced and xsave runs again this does not remove the previously saved state from
the xsave region or the XSTATE_BV field in the header. Performing an fxrstor then panics
as XSTATE_BV has state bits that are not set in xcr0.

Two potential alternate solutions for fixing this are
 1. Manually removing now unused state bits from XSTATE_BV
 2. Make guest xcr0 bits 'sticky' and once enabled keep them enabled

Both of these solutions are more complex and compete with existing hardware optimizations
of not saving and restoring state that hasn't changed.

Test: hypervisor unit tests + guest integration tests on NUC + qemu

Change-Id: I8c6b3e972800fde48c0aa6283c60d8796bf864aa