[tee][optee] Bring in ta_storage.h from optee_test
This change adds ta_storage.h from the optee_test suite. This defines
the values expected by the storage test trusted application, and can be
used in a client application to test OP-TEE RPC compatibility on
Fuchsia.

Test: No behavioral change. This is just a third party header from the
optee_test suite that is needed for client-side code to properly
interface with a certain application in the trusted world.

OSRB-175 #done

Change-Id: I0439aa8a05d92d0ac40e72b09ce3c07088625885