[futex][trace] Add ktrace events for futex ops. (kernel)
Add ktrace events for the 4 main futex operations.  Default them to be
off at compile time for now.

Note: this CL only adds trace definitions and tracing behavior to the
kernel.  To make any use of these events, a follow up change which
implementes the trace importer behavior is needed.

Change-Id: Ia31d42234ac1cda1d66091cf56e7a8447e58facb