[mtk-power] Do not fail if the rail is already enabled
When enabling power domain, if the rail is already enabled, we fail.
This changeset rectifies this wrong behavior by returning success.
All this code is changing with https://fuchsia-review.googlesource.com/c/fuchsia/+/273632/
and that change will not be submitted without the unit tests, that I am
currently writing. The current change is to unblock folks before the
other change is reviewed.

ZX-3990 #Done

Test: fx full-build. Without this change, I see the bind failure for
power driver. With this change, I do not see the error. device-enumeration-test is fixed here
https://fuchsia-review.googlesource.com/c/fuchsia/+/277336. It fails without this change.
Succeeds with this change

Change-Id: Iec6fe6d8ec041a301287e4f0b787c3b5519d47d6