[pseudo-fs] [file::asynchronous] Remove '+ Unpin'
I have spent quite some time on this, and if we want full support for
`Pin`ed closures, we would need to add a bit of code.  But as this issue
is currently only affecting the async callbacks, this change addresses
just that part of the code.

I do not know if there is a useful `FnMut` that does not implement
`Unpin`.  So far we have seen it for the `Future`s and `Stream`s only.

ZX-3913 #done

Change-Id: I4ab6d33a1d6b64ffb5704f75905df6f6592a772e