[debugger] Attach to the root job by default.
Previously the debugger would automatically attach to the component
root. But this misses some processes. What came up recently was Zircon
system tests. Having to do extra steps to watch the right job for these
common tests is confusing for developers.

Now the system root job is attached to by default on connection.

Added switches to the "attach" verb to refer to either the system root
or component root job without having to look up its koid.

Enhanced the "attach" documentation.

Did some minor cleanup of the job status messaging. The state was copied
from the process object and included a state we didn't use "starting"
and using "running" for a job is confusing. The messaging is now either
"Not attached", "Attaching", or "Attached".

Change-Id: I386921f3efb4b310b36dfd27c8773e86d743e6e7